+++ apps/3d_rec_framework/include/pcl/apps/3d_rec_framework/feature_wrapper/global/ourcvfh_estimator.h
@@ -25,12 +25,12 @@ namespace pcl
         typedef typename pcl::PointCloud<PointInT>::Ptr PointInTPtr;
         using GlobalEstimator<PointInT, FeatureT>::normal_estimator_;
         using GlobalEstimator<PointInT, FeatureT>::normals_;
-        float eps_angle_threshold_;
-        float curvature_threshold_;
-        float cluster_tolerance_factor_;
+        double eps_angle_threshold_;
+        double curvature_threshold_;
+        double cluster_tolerance_factor_;
         bool normalize_bins_;
         bool adaptative_MLS_;
-        float refine_factor_;
+        double refine_factor_;
 
         std::vector<bool> valid_roll_transforms_;
         std::vector<Eigen::Matrix4f, Eigen::aligned_allocator<Eigen::Matrix4f> > transforms_;
@@ -48,7 +48,7 @@ namespace pcl
         }
 
         void
-        setCVFHParams (float p1, float p2, float p3)
+        setCVFHParams (double p1, double p2, double p3)
         {
           eps_angle_threshold_ = p1;
           curvature_threshold_ = p2;
@@ -62,7 +62,7 @@ namespace pcl
         }
 
         void
-        setRefineClustersParam (float p4)
+        setRefineClustersParam (double p4)
         {
           refine_factor_ = p4;
         }
@@ -99,8 +99,8 @@ namespace pcl
             typename search::KdTree<PointInT>::Ptr tree;
             Eigen::Vector4f centroid_cluster;
             pcl::compute3DCentroid (*in, centroid_cluster);
-            float dist_to_sensor = centroid_cluster.norm ();
-            float sigma = dist_to_sensor * 0.01f;
+            double dist_to_sensor = centroid_cluster.norm ();
+            double sigma = dist_to_sensor * 0.01f;
             mls.setSearchMethod (tree);
             mls.setSearchRadius (sigma);
             mls.setUpsamplingMethod (mls.SAMPLE_LOCAL_PLANE);
@@ -144,8 +144,8 @@ namespace pcl
           cvfh.setNormalizeBins (normalize_bins_);
           cvfh.setRefineClusters(refine_factor_);
 
-          float radius = normal_estimator_->normal_radius_;
-          float cluster_tolerance_radius = normal_estimator_->grid_resolution_ * cluster_tolerance_factor_;
+          double radius = normal_estimator_->normal_radius_;
+          double cluster_tolerance_radius = normal_estimator_->grid_resolution_ * cluster_tolerance_factor_;
 
           if (normal_estimator_->compute_mesh_resolution_)
           {
