+++ apps/3d_rec_framework/include/pcl/apps/3d_rec_framework/feature_wrapper/local/shot_local_estimator_omp.h
@@ -55,8 +55,8 @@ namespace pcl
             typename search::KdTree<PointInT>::Ptr tree;
             Eigen::Vector4f centroid_cluster;
             pcl::compute3DCentroid (*in, centroid_cluster);
-            float dist_to_sensor = centroid_cluster.norm ();
-            float sigma = dist_to_sensor * 0.01f;
+            double dist_to_sensor = centroid_cluster.norm ();
+            double sigma = dist_to_sensor * 0.01f;
             mls.setSearchMethod (tree);
             mls.setSearchRadius (sigma);
             mls.setUpsamplingMethod (mls.SAMPLE_LOCAL_PLANE);
@@ -118,7 +118,7 @@ namespace pcl
           signatures->width = static_cast<int> (shots->points.size ());
           signatures->height = 1;
 
-          int size_feat = sizeof(signatures->points[0].histogram) / sizeof(float);
+          int size_feat = sizeof(signatures->points[0].histogram) / sizeof(double);
 
           int good = 0;
           for (size_t k = 0; k < shots->points.size (); k++)
